# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_javinizer_javinizer-go_internal_api_contracts.ActressView:
    properties:
      aliases:
        type: string
      created_at:
        type: string
      dmm_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      last_name:
        type: string
      name_key:
        type: string
      origin:
        type: string
      thumb_url:
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ActressTranslationView'
        type: array
      updated_at:
        type: string
      verified:
        type: boolean
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeFailed:
    properties:
      error:
        example: Result not found in job
        type: string
      result_id:
        example: uuid-1
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeRequest:
    properties:
      result_ids:
        example:
        - uuid-1
        - uuid-2
        items:
          type: string
        type: array
    required:
    - result_ids
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeResponse:
    properties:
      excluded:
        items:
          type: string
        type: array
      failed:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeFailed'
        type: array
      job:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse'
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchFileResult:
    properties:
      actress_sources:
        additionalProperties:
          type: string
        type: object
      ended_at:
        type: string
      error:
        type: string
      error_code:
        type: string
      field_sources:
        additionalProperties:
          type: string
        type: object
      file_path:
        type: string
      is_multi_part:
        type: boolean
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
      movie_id:
        type: string
      part_number:
        type: integer
      part_suffix:
        type: string
      result_id:
        description: Stable UUID — survives movie_id changes
        type: string
      revision:
        description: |-
          Revision: server-managed update counter used as the CAS baseline for
          PATCH expected_result_revision (POSTER-WRITE-HARDENING D12).
        type: integer
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus'
      translation_warning:
        description: |-
          TranslationWarning is populated when metadata translation partially failed
          or degraded; empty when translation fully succeeded.
        type: string
      translation_warning_code:
        description: |-
          TranslationWarningCode is the machine-readable classification of
          TranslationWarning (rate_limited, unavailable, degraded, ...).
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse:
    properties:
      apply_generation:
        type: integer
      apply_plan:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.Plan'
      completed:
        type: integer
      completed_at:
        type: string
      destination:
        type: string
      excluded:
        additionalProperties:
          type: boolean
        type: object
      failed:
        type: integer
      files:
        items:
          type: string
        type: array
      id:
        type: string
      noop_count:
        description: |-
          NoopCount: terminal completed-noop rows (authorized duplicate skips —
          codex P2, PR #241 F2). Non-revertible; revertible = operation_count −
          reverted_count − noop_count.
        type: integer
      operation_count:
        type: integer
      operation_mode_override:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_operationmode.OperationMode'
      persist_error:
        type: string
      progress:
        type: number
      results:
        additionalProperties:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchFileResult'
        type: object
      reverted_count:
        type: integer
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus'
      total_files:
        type: integer
      update:
        type: boolean
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.DataSource:
    properties:
      confidence:
        description: Confidence score (0.0-1.0)
        example: 0.9
        type: number
      last_updated:
        description: ISO 8601 timestamp
        example: "2024-01-15T10:30:00Z"
        type: string
      source:
        description: '"scraper" or "nfo"'
        example: nfo
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.DisplayTitlePreviewRequest:
    properties:
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.DisplayTitlePreviewResponse:
    properties:
      display_title:
        example: '[IPX-123] Beautiful Woman'
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse:
    properties:
      code:
        description: Stable machine-readable error code (e.g. "JOB_NOT_FOUND") for
          client-side i18n
        type: string
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
      params:
        additionalProperties: {}
        description: Structured arguments for translating Code into a localized message
        type: object
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.GenreView:
    properties:
      id:
        type: integer
      name:
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.GenreTranslationView'
        type: array
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.MergeStatistics:
    properties:
      conflicts_resolved:
        example: 5
        type: integer
      empty_fields:
        example: 2
        type: integer
      from_nfo:
        example: 3
        type: integer
      from_scraper:
        example: 10
        type: integer
      merged_arrays:
        example: 2
        type: integer
      total_fields:
        example: 15
        type: integer
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieResponse:
    properties:
      merge_stats:
        allOf:
        - $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MergeStatistics'
        description: Merge statistics when NFO merging occurred
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
      provenance:
        additionalProperties:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.DataSource'
        description: Field-level data source tracking
        type: object
      revision:
        description: Revision is the fresh post-commit result revision (POSTER-WRITE-HARDENING
          D12).
        type: integer
      revisions:
        additionalProperties:
          format: int64
          type: integer
        description: |-
          Revisions carries EVERY family part's fresh revision keyed by
          result_id (multipart-safe CAS baselines, codex r26).
        type: object
      translation_warning:
        description: |-
          TranslationWarning is populated on scrape-like flows (rescrape) when
          metadata translation partially failed or degraded.
        type: string
      translation_warning_code:
        description: |-
          TranslationWarningCode is the machine-readable classification of
          TranslationWarning (rate_limited, unauthorized, forbidden, request_error,
          service_error, unavailable, degraded, unknown).
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieTranslationView:
    properties:
      created_at:
        type: string
      description:
        type: string
      director:
        type: string
      id:
        type: integer
      label:
        type: string
      language:
        type: string
      maker:
        type: string
      movie_id:
        type: string
      original_title:
        type: string
      series:
        type: string
      settings_hash:
        type: string
      source_name:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieView:
    properties:
      actresses:
        description: |-
          Relationships (contract DTOs — see collection_views.go; persistence
          models.* types never cross the API boundary)
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ActressView'
        type: array
      cast_version:
        type: string
      code:
        description: Canonical JAV code (e.g., "IPX-535"). Renamed from content_id.
        example: IPX-535
        type: string
      cover_url:
        type: string
      created_at:
        description: Audit timestamps
        type: string
      cropped_poster_url:
        type: string
      description:
        description: Metadata
        type: string
      director:
        type: string
      display_title:
        description: Titles
        example: IPX-535 Beautiful Woman
        type: string
      genres:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.GenreView'
        type: array
      id:
        description: Secondary identifier — often equals Code but can differ.
        example: IPX-535
        type: string
      label:
        description: Sub-label
        type: string
      maker:
        description: Studio/maker
        type: string
      original_cover_url:
        type: string
      original_cropped_poster_url:
        type: string
      original_filename:
        type: string
      original_poster_url:
        type: string
      original_should_crop_poster:
        type: boolean
      original_title:
        example: 美人
        type: string
      poster_crop_bounds:
        allOf:
        - $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.CropBounds'
        description: Poster / cover (flattened from PosterState — no custom marshaler
          needed)
      poster_crop_source_full:
        type: boolean
      poster_url:
        type: string
      rating_score:
        type: number
      rating_votes:
        type: integer
      rating_warning:
        type: string
      release_date:
        type: string
      release_year:
        type: integer
      runtime:
        description: in minutes
        type: integer
      screenshot_urls:
        items:
          type: string
        type: array
      series:
        description: Series name
        type: string
      should_crop_poster:
        type: boolean
      source_name:
        description: Source provenance
        type: string
      source_url:
        type: string
      title:
        example: Beautiful Woman
        type: string
      trailer_url:
        description: Media
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieTranslationView'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.UpdateMovieRequest:
    properties:
      expected_result_revision:
        description: |-
          ExpectedResultRevision (POSTER-WRITE-HARDENING D12): optional CAS
          guard — when set, the save 409s if the target result's current
          revision differs, so a stale-snapshot client cannot silently clobber a
          newer committed edit.
        type: integer
      expected_result_revisions:
        additionalProperties:
          format: int64
          type: integer
        description: |-
          ExpectedResultRevisions: per-part CAS for multipart families (codex
          r39) - EVERY listed result must currently sit at the mapped revision or
          the whole save 409s before any write.
        type: object
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
    required:
    - movie
    type: object
...
  github_com_javinizer_javinizer-go_internal_applyplan.Plan:
    properties:
      destination:
        type: string
      media_policy:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.MediaPolicy'
      merge:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.MergePolicy'
      nfo_output:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.NFOOutput'
      version:
        type: integer
      video_operation:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_applyplan.VideoOperation'
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.CropBounds:
    properties:
      height:
        type: number
      source_aspect:
        description: |-
          SourceAspect is the width/height ratio of the source image the crop was
          measured against. The apply phase refuses geometry whose aspect no
          longer matches the downloaded image, so same-URL source swaps fall back
          to pre-change behavior instead of cropping the wrong image.
        type: number
      source_fingerprint:
        description: |-
          SourceFingerprint binds the normalized geometry to the exact full-size
          bytes measured in the review UI. Empty is retained for legacy envelopes;
          non-empty values must be a canonical SHA-256 hex digest.
        type: string
      width:
        type: number
      x:
        type: number
      "y":
        type: number
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.JobStatus:
    enum:
    - pending
    - running
    - completed
    - failed
    - cancelled
    - organized
    - reverted
    type: string
    x-enum-varnames:
    - JobStatusPending
    - JobStatusRunning
    - JobStatusCompleted
    - JobStatusFailed
    - JobStatusCancelled
    - JobStatusOrganized
    - JobStatusReverted
...
  github_com_javinizer_javinizer-go_internal_operationmode.OperationMode:
    enum:
    - organize
    - in-place
    - in-place-norenamefolder
    - metadata-artwork
    - preview
    type: string
    x-enum-varnames:
    - OperationModeOrganize
    - OperationModeInPlace
    - OperationModeInPlaceNoRenameFolder
    - OperationModeMetadataArtwork
    - OperationModePreview
...
host: localhost:8765
info:
  contact:
    name: API Support
    url: https://github.com/javinizer/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/javinizer/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/batch/{id}/movies/batch-exclude:
    post:
      consumes:
      - application/json
      description: 'Exclude multiple movies from a batch job in a single request.
        Best-effort: excludes as many as possible and returns per-result failures.'
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result IDs to exclude
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "409":
          description: a phase is running
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "410":
          description: job deleted
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Bulk exclude movies from batch organization
      tags:
      - web
...
  /api/v1/batch/{id}/results/{resultId}:
    patch:
      consumes:
      - application/json
      description: Update a movie's metadata within a batch job's results
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      - description: Updated movie data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.UpdateMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "409":
          description: job busy (pending or scrape-phase)
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "410":
          description: job deleted
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: transactional save failed; all writes rolled back
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Update movie in batch job
      tags:
      - web
  /api/v1/batch/{id}/results/{resultId}/display-title-preview:
    post:
      consumes:
      - application/json
      description: Render the configured display_title template for the provided movie
        using the shared workflow template engine, without mutating persisted state.
        Used by the review page to show a live NFO title preview as the user edits
        the base title.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      - description: Edited movie to render
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.DisplayTitlePreviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.DisplayTitlePreviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Preview the rendered display_title template
      tags:
      - web
  /api/v1/batch/{id}/results/{resultId}/exclude:
    post:
      description: Mark a movie in a batch job as excluded from file organization
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "409":
          description: a phase is running
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "410":
          description: job deleted
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Exclude movie from batch organization
      tags:
      - web
...
schemes:
- http
- https
swagger: "2.0"